requirements.txt
pdfminer.six
python-docx
spacy
easyocr
Pillow